backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
textblob==0.17.1
google-generativeai==0.3.1
python-dotenv==1.0.0
nltk==3.8.1